refactor : 로그인 함수 분리(#33)
